+++ CHANGELOG.md
@@ -6,4 +6,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
-- Initial release
+### Fixed
+
+- v0.0.1 Errors in stopping server
+- v0.0.1 Now the extension does not need the cursor to be on the file before starting